tests(all): refactor all tests and add missing
