+++ packages/bridge/tests/events/onTelegramEvent.ts
@@ -1,8 +1,8 @@
 import { expect, vi, beforeEach, afterEach, describe, it } from 'vitest';
 
 import { onTelegramEvent } from '../../src/events/onTelegramEvent.js';
-import { createWindow, type WindowSpy } from '../../__test-utils__/createWindow.js';
-import { dispatchWindowMessageEvent } from '../../__test-utils__/dispatchWindowMessageEvent.js';
+import { createWindow, type WindowSpy } from '../../test-utils/createWindow.js';
+import { dispatchWindowMessageEvent } from '../../test-utils/dispatchWindowMessageEvent.js';
 
 let windowSpy: WindowSpy;
