+++ packages/bridge/tests/events/subscribe.ts
@@ -1,8 +1,8 @@
 import { expect, describe, it, vi, beforeEach, afterEach } from 'vitest';
 
 import { subscribe } from '../../src/index.js';
-import { createWindow, type WindowSpy } from '../../__test-utils__/createWindow.js';
-import { dispatchWindowMessageEvent } from '../../__test-utils__/dispatchWindowMessageEvent.js';
+import { createWindow, type WindowSpy } from '../../test-utils/createWindow.js';
+import { dispatchWindowMessageEvent } from '../../test-utils/dispatchWindowMessageEvent.js';
 
 let windowSpy: WindowSpy;
